+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Replaced usage of the deprecated io/ioutil package.
 - Audit /download_results endpoint support
+- Sanitize service support
 
 ### Fixed